+++ cppcrypto/kalyna.h
@@ -15,8 +15,8 @@ namespace cppcrypto
 	class kalyna512_512 : public block_cipher
 	{
 	public:
-		int blocksize() const { return 512; }
-		int keysize() const { return 512; }
+		size_t blocksize() const { return 512; }
+		size_t keysize() const { return 512; }
 		block_cipher* clone() const { return new kalyna512_512; }
 		void clear();
 
@@ -31,8 +31,8 @@ namespace cppcrypto
 	class kalyna256_512 : public block_cipher
 	{
 	public:
-		int blocksize() const { return 256; }
-		int keysize() const { return 512; }
+		size_t blocksize() const { return 256; }
+		size_t keysize() const { return 512; }
 		block_cipher* clone() const { return new kalyna256_512; }
 		void clear();
 
@@ -47,8 +47,8 @@ namespace cppcrypto
 	class kalyna256_256 : public block_cipher
 	{
 	public:
-		int blocksize() const { return 256; }
-		int keysize() const { return 256; }
+		size_t blocksize() const { return 256; }
+		size_t keysize() const { return 256; }
 		block_cipher* clone() const { return new kalyna256_256; }
 		void clear();
 
@@ -63,8 +63,8 @@ namespace cppcrypto
 	class kalyna128_256 : public block_cipher
 	{
 	public:
-		int blocksize() const { return 128; }
-		int keysize() const { return 256; }
+		size_t blocksize() const { return 128; }
+		size_t keysize() const { return 256; }
 		block_cipher* clone() const { return new kalyna128_256; }
 		void clear();
 
@@ -79,8 +79,8 @@ namespace cppcrypto
 	class kalyna128_128 : public block_cipher
 	{
 	public:
-		int blocksize() const { return 128; }
-		int keysize() const { return 128; }
+		size_t blocksize() const { return 128; }
+		size_t keysize() const { return 128; }
 		block_cipher* clone() const { return new kalyna128_128; }
 		void clear();
 
@@ -96,3 +96,4 @@ namespace cppcrypto
 }
 
 #endif
+
