+++ cppcrypto/rijndael.h
@@ -20,8 +20,8 @@ namespace cppcrypto
 		rijndael128_128();
 		~rijndael128_128();
 
-		int blocksize() const { return 128; }
-		int keysize() const { return 128; }
+		size_t blocksize() const { return 128; }
+		size_t keysize() const { return 128; }
 		block_cipher* clone() const { return new rijndael128_128; }
 		void clear();
 
@@ -40,8 +40,8 @@ namespace cppcrypto
 		rijndael128_160();
 		~rijndael128_160();
 
-		int blocksize() const { return 128; }
-		int keysize() const { return 160; }
+		size_t blocksize() const { return 128; }
+		size_t keysize() const { return 160; }
 		block_cipher* clone() const { return new rijndael128_160; }
 		void clear();
 
@@ -60,8 +60,8 @@ namespace cppcrypto
 		rijndael128_192();
 		~rijndael128_192();
 
-		int blocksize() const { return 128; }
-		int keysize() const { return 192; }
+		size_t blocksize() const { return 128; }
+		size_t keysize() const { return 192; }
 		block_cipher* clone() const { return new rijndael128_192; }
 		void clear();
 
@@ -80,8 +80,8 @@ namespace cppcrypto
 		rijndael128_224();
 		~rijndael128_224();
 
-		int blocksize() const { return 128; }
-		int keysize() const { return 224; }
+		size_t blocksize() const { return 128; }
+		size_t keysize() const { return 224; }
 		block_cipher* clone() const { return new rijndael128_224; }
 		void clear();
 
@@ -100,8 +100,8 @@ namespace cppcrypto
 		rijndael128_256();
 		~rijndael128_256();
 
-		int blocksize() const { return 128; }
-		int keysize() const { return 256; }
+		size_t blocksize() const { return 128; }
+		size_t keysize() const { return 256; }
 		block_cipher* clone() const { return new rijndael128_256; }
 		void clear();
 
@@ -123,7 +123,7 @@ namespace cppcrypto
 			~rijndael256();
 			void clear();
 
-			int blocksize() const { return 256; }
+			size_t blocksize() const { return 256; }
 
 			void encrypt_block(const uint8_t* in, uint8_t* out);
 			void decrypt_block(const uint8_t* in, uint8_t* out);
@@ -139,7 +139,7 @@ namespace cppcrypto
 	public:
 		rijndael256_128();
 
-		int keysize() const { return 128; }
+		size_t keysize() const { return 128; }
 		block_cipher* clone() const { return new rijndael256_128; }
 
 		bool init(const uint8_t* key, block_cipher::direction direction);
@@ -150,7 +150,7 @@ namespace cppcrypto
 	public:
 		rijndael256_160();
 
-		int keysize() const { return 160; }
+		size_t keysize() const { return 160; }
 		block_cipher* clone() const { return new rijndael256_160; }
 
 		bool init(const uint8_t* key, block_cipher::direction direction);
@@ -161,7 +161,7 @@ namespace cppcrypto
 	public:
 		rijndael256_192();
 
-		int keysize() const { return 192; }
+		size_t keysize() const { return 192; }
 		block_cipher* clone() const { return new rijndael256_192; }
 
 		bool init(const uint8_t* key, block_cipher::direction direction);
@@ -172,7 +172,7 @@ namespace cppcrypto
 	public:
 		rijndael256_224();
 
-		int keysize() const { return 224; }
+		size_t keysize() const { return 224; }
 		block_cipher* clone() const { return new rijndael256_224; }
 
 		bool init(const uint8_t* key, block_cipher::direction direction);
@@ -183,7 +183,7 @@ namespace cppcrypto
 	public:
 		rijndael256_256();
 
-		int keysize() const { return 256; }
+		size_t keysize() const { return 256; }
 		block_cipher* clone() const { return new rijndael256_256; }
 
 		bool init(const uint8_t* key, block_cipher::direction direction);
@@ -195,8 +195,8 @@ namespace cppcrypto
 		rijndael192_128();
 		~rijndael192_128();
 
-		int blocksize() const { return 192; }
-		int keysize() const { return 128; }
+		size_t blocksize() const { return 192; }
+		size_t keysize() const { return 128; }
 		block_cipher* clone() const { return new rijndael192_128; }
 		void clear();
 
@@ -215,8 +215,8 @@ namespace cppcrypto
 		rijndael192_160();
 		~rijndael192_160();
 
-		int blocksize() const { return 192; }
-		int keysize() const { return 160; }
+		size_t blocksize() const { return 192; }
+		size_t keysize() const { return 160; }
 		block_cipher* clone() const { return new rijndael192_160; }
 		void clear();
 
@@ -236,8 +236,8 @@ namespace cppcrypto
 		rijndael192_192();
 		~rijndael192_192();
 
-		int blocksize() const { return 192; }
-		int keysize() const { return 192; }
+		size_t blocksize() const { return 192; }
+		size_t keysize() const { return 192; }
 		block_cipher* clone() const { return new rijndael192_192; }
 		void clear();
 
@@ -256,8 +256,8 @@ namespace cppcrypto
 		rijndael192_224();
 		~rijndael192_224();
 
-		int blocksize() const { return 192; }
-		int keysize() const { return 224; }
+		size_t blocksize() const { return 192; }
+		size_t keysize() const { return 224; }
 		block_cipher* clone() const { return new rijndael192_224; }
 		void clear();
 
@@ -276,8 +276,8 @@ namespace cppcrypto
 		rijndael192_256();
 		~rijndael192_256();
 
-		int blocksize() const { return 192; }
-		int keysize() const { return 256; }
+		size_t blocksize() const { return 192; }
+		size_t keysize() const { return 256; }
 		block_cipher* clone() const { return new rijndael192_256; }
 		void clear();
 
@@ -296,8 +296,8 @@ namespace cppcrypto
 		rijndael160_128();
 		~rijndael160_128();
 
-		int blocksize() const { return 160; }
-		int keysize() const { return 128; }
+		size_t blocksize() const { return 160; }
+		size_t keysize() const { return 128; }
 		block_cipher* clone() const { return new rijndael160_128; }
 		void clear();
 
@@ -316,8 +316,8 @@ namespace cppcrypto
 		rijndael160_160();
 		~rijndael160_160();
 
-		int blocksize() const { return 160; }
-		int keysize() const { return 160; }
+		size_t blocksize() const { return 160; }
+		size_t keysize() const { return 160; }
 		block_cipher* clone() const { return new rijndael160_160; }
 		void clear();
 
@@ -336,8 +336,8 @@ namespace cppcrypto
 		rijndael160_192();
 		~rijndael160_192();
 
-		int blocksize() const { return 160; }
-		int keysize() const { return 192; }
+		size_t blocksize() const { return 160; }
+		size_t keysize() const { return 192; }
 		block_cipher* clone() const { return new rijndael160_192; }
 		void clear();
 
@@ -356,8 +356,8 @@ namespace cppcrypto
 		rijndael160_224();
 		~rijndael160_224();
 
-		int blocksize() const { return 160; }
-		int keysize() const { return 224; }
+		size_t blocksize() const { return 160; }
+		size_t keysize() const { return 224; }
 		block_cipher* clone() const { return new rijndael160_224; }
 		void clear();
 
@@ -376,8 +376,8 @@ namespace cppcrypto
 		rijndael160_256();
 		~rijndael160_256();
 
-		int blocksize() const { return 160; }
-		int keysize() const { return 256; }
+		size_t blocksize() const { return 160; }
+		size_t keysize() const { return 256; }
 		block_cipher* clone() const { return new rijndael160_256; }
 		void clear();
 
@@ -400,7 +400,7 @@ namespace cppcrypto
 			~rijndael224();
 			void clear();
 
-			int blocksize() const { return 224; }
+			size_t blocksize() const { return 224; }
 
 			void encrypt_block(const uint8_t* in, uint8_t* out);
 			void decrypt_block(const uint8_t* in, uint8_t* out);
@@ -415,7 +415,7 @@ namespace cppcrypto
 	{
 	public:
 		rijndael224_128();
-		int keysize() const { return 128; }
+		size_t keysize() const { return 128; }
 		block_cipher* clone() const { return new rijndael224_128; }
 
 		bool init(const uint8_t* key, block_cipher::direction direction);
@@ -425,7 +425,7 @@ namespace cppcrypto
 	{
 	public:
 		rijndael224_160();
-		int keysize() const { return 160; }
+		size_t keysize() const { return 160; }
 		block_cipher* clone() const { return new rijndael224_160; }
 
 		bool init(const uint8_t* key, block_cipher::direction direction);
@@ -435,7 +435,7 @@ namespace cppcrypto
 	{
 	public:
 		rijndael224_192();
-		int keysize() const { return 192; }
+		size_t keysize() const { return 192; }
 		block_cipher* clone() const { return new rijndael224_192; }
 
 		bool init(const uint8_t* key, block_cipher::direction direction);
@@ -445,7 +445,7 @@ namespace cppcrypto
 	{
 	public:
 		rijndael224_224();
-		int keysize() const { return 224; }
+		size_t keysize() const { return 224; }
 		block_cipher* clone() const { return new rijndael224_224; }
 
 		bool init(const uint8_t* key, block_cipher::direction direction);
@@ -457,8 +457,8 @@ namespace cppcrypto
 		rijndael224_256();
 		~rijndael224_256();
 
-		int blocksize() const { return 224; }
-		int keysize() const { return 256; }
+		size_t blocksize() const { return 224; }
+		size_t keysize() const { return 256; }
 		block_cipher* clone() const { return new rijndael224_256; }
 		void clear();
