+++ cppcrypto/twofish.h
@@ -17,7 +17,7 @@ namespace cppcrypto
 		class twofish : public block_cipher
 		{
 		public:
-			int blocksize() const { return 128; }
+			size_t blocksize() const { return 128; }
 			void encrypt_block(const uint8_t* in, uint8_t* out);
 			void decrypt_block(const uint8_t* in, uint8_t* out);
 			void clear();
@@ -31,7 +31,7 @@ namespace cppcrypto
 	class twofish128 : public detail::twofish
 	{
 	public:
-		int keysize() const { return 128; }
+		size_t keysize() const { return 128; }
 		block_cipher* clone() const { return new twofish128; }
 
 		bool init(const uint8_t* key, block_cipher::direction direction);
@@ -40,7 +40,7 @@ namespace cppcrypto
 	class twofish192 : public detail::twofish
 	{
 	public:
-		int keysize() const { return 192; }
+		size_t keysize() const { return 192; }
 		block_cipher* clone() const { return new twofish192; }
 
 		bool init(const uint8_t* key, block_cipher::direction direction);
@@ -49,7 +49,7 @@ namespace cppcrypto
 	class twofish256 : public detail::twofish
 	{
 	public:
-		int keysize() const { return 256; }
+		size_t keysize() const { return 256; }
 		block_cipher* clone() const { return new twofish256; }
 
 		bool init(const uint8_t* key, block_cipher::direction direction);
